test: address review comments and improve Windows compatibility
- Move permission tests (os.Getuid/os.Chmod) to _unix_test.go with build tag
- Remove tautological tests that only re-check struct field values
- Use filepath.Join for all path config values instead of forward slashes
- Add t.Cleanup for backend integration test fixture file
- Fix godot linter comments
